add test for get proof
