+++ core/src/test/resources/testGetProofForAvailableKey.json
@@ -1,0 +1,51 @@
+{
+  "key":"0x0000000000000000000000000000000000000000000000000000000000000024",
+  "leafIndex":2,
+  "proof":{
+    "value":"0x0000000000000000000000000000000000000000000000000000000000000020",
+    "proofRelatedNodes":[
+      "0x0000000000000000000000000000000000000000000000000000000000000003065616ed1369e9992cd4b19806b58547443c3856a695deb8494f127782e7b38d",
+      "0x1254d0207708e41336c9e607dfb109ed26eaa974b75ff9ababcfde3ff70a169a1254d0207708e41336c9e607dfb109ed26eaa974b75ff9ababcfde3ff70a169a",
+      "0x1798747385b7e3752743a21e1dde652c9dee1106d61fa56ccb91cff1e64348841798747385b7e3752743a21e1dde652c9dee1106d61fa56ccb91cff1e6434884",
+      "0x1c477aa6e33b02996c81cedf8d1e88c9cc6ea40784191e391d23979dcb0208ec1c477aa6e33b02996c81cedf8d1e88c9cc6ea40784191e391d23979dcb0208ec",
+      "0x2d3dbe9694f65a1a380e3cdc92145a6e4815ce22304dab5ec90207f0493c8a212d3dbe9694f65a1a380e3cdc92145a6e4815ce22304dab5ec90207f0493c8a21",
+      "0x190bec88beb0a878950c6d48f5e9a5b84f65a699a48e9b176152abee95ae2cf6190bec88beb0a878950c6d48f5e9a5b84f65a699a48e9b176152abee95ae2cf6",
+      "0x23f079e1eb77786681e1447bc5ddb2b0ea7ccb01e40a99b6ff56339c884ba13d23f079e1eb77786681e1447bc5ddb2b0ea7ccb01e40a99b6ff56339c884ba13d",
+      "0x01abc4d0808e21dc4755d7e39a001b9efd6fec73af45bbc09fffaf6a7035c08101abc4d0808e21dc4755d7e39a001b9efd6fec73af45bbc09fffaf6a7035c081",
+      "0x0e6109c81f60a9a1dc1a1dc417b8d61f8b005c4bd18f2f26d35d29d92e90c7fb0e6109c81f60a9a1dc1a1dc417b8d61f8b005c4bd18f2f26d35d29d92e90c7fb",
+      "0x168961f185851b92bb9f7e130f872c65c2ef2e78e28d0b7f1d65c1e88db3506b168961f185851b92bb9f7e130f872c65c2ef2e78e28d0b7f1d65c1e88db3506b",
+      "0x15fe20db6b9027a70037d96311e02f9743931c857747bad2d969b3a84fc52e5515fe20db6b9027a70037d96311e02f9743931c857747bad2d969b3a84fc52e55",
+      "0x298c7e7a54599edb03e2a5d4336245696e9b181f5d6929d8869516f319c0e2b0298c7e7a54599edb03e2a5d4336245696e9b181f5d6929d8869516f319c0e2b0",
+      "0x255e9ebb837d0e1f8f319d3d479294abe31254aedc63250e5d7db32e23d8ef9d255e9ebb837d0e1f8f319d3d479294abe31254aedc63250e5d7db32e23d8ef9d",
+      "0x28063ba1db95bda8a256f10fc6a74b0be7e54bbb2971ec45451b903ce872f0e828063ba1db95bda8a256f10fc6a74b0be7e54bbb2971ec45451b903ce872f0e8",
+      "0x22fe5e952f220e4f2ca8ff87e5921b08683064c521fea942bf92c8e2532f92ff22fe5e952f220e4f2ca8ff87e5921b08683064c521fea942bf92c8e2532f92ff",
+      "0x02c296e675a4c76f85779386952762c8c15020c90531190bd2c20c2f522e491d02c296e675a4c76f85779386952762c8c15020c90531190bd2c20c2f522e491d",
+      "0x1f4a48ef184635ec978659c1f5bbb3b633b52ac1cfdf03327958f305259dfe621f4a48ef184635ec978659c1f5bbb3b633b52ac1cfdf03327958f305259dfe62",
+      "0x1f44b7c7fdc4efebb3799352c5b7fe449d62d5e6e0fb7297b328c7e2b3d7eef01f44b7c7fdc4efebb3799352c5b7fe449d62d5e6e0fb7297b328c7e2b3d7eef0",
+      "0x01987180f8ab0ad6da96c2bd4ad93b28a58ed8fc020b85127f156a603d3569be01987180f8ab0ad6da96c2bd4ad93b28a58ed8fc020b85127f156a603d3569be",
+      "0x19980a40aff0d7157f11ea2e834116494721ccfbaa9d90d6b1b69a66621f555219980a40aff0d7157f11ea2e834116494721ccfbaa9d90d6b1b69a66621f5552",
+      "0x253c6498543a14e1bfdd21b2c5ea572666ba6c450075b6c104294dcfbb5a29c3253c6498543a14e1bfdd21b2c5ea572666ba6c450075b6c104294dcfbb5a29c3",
+      "0x12354fbc87fb38377cbfbeae87585eb141c373aa61a072ff86cec7efd5c7a56912354fbc87fb38377cbfbeae87585eb141c373aa61a072ff86cec7efd5c7a569",
+      "0x24dbf0f1835370bf7f9ec5eb500bbeac2bf5c41504f29999cddf3c79eaa9bf3724dbf0f1835370bf7f9ec5eb500bbeac2bf5c41504f29999cddf3c79eaa9bf37",
+      "0x2b44e255c63ae690467e1541623c37ac9e2d06bb95b794debbb40dc69973a4452b44e255c63ae690467e1541623c37ac9e2d06bb95b794debbb40dc69973a445",
+      "0x0b36746bc15c76db06d9b0a3b399ce9372403f1b928431dd544059238e9e1d150b36746bc15c76db06d9b0a3b399ce9372403f1b928431dd544059238e9e1d15",
+      "0x0e6038b3d33f66cf8bd05e0511862e165ce79afbd2b03a6d6a0a0818148316bd0e6038b3d33f66cf8bd05e0511862e165ce79afbd2b03a6d6a0a0818148316bd",
+      "0x26afa3a82762a461a8895be815960a9a61acdd641223176789263c79f20654eb26afa3a82762a461a8895be815960a9a61acdd641223176789263c79f20654eb",
+      "0x0df1496ebc60b806f4f712155ae3953493b54d911386b03033a075d8550813e10df1496ebc60b806f4f712155ae3953493b54d911386b03033a075d8550813e1",
+      "0x1c26c9e8963ed1d3e8e18eb36672186cfbda944795465db9496780a13e0b082d1c26c9e8963ed1d3e8e18eb36672186cfbda944795465db9496780a13e0b082d",
+      "0x214dc26f4b6f38c484cb304fdf042bc7b9402640a2f1fc6c7cfb1b006c17c390214dc26f4b6f38c484cb304fdf042bc7b9402640a2f1fc6c7cfb1b006c17c390",
+      "0x0ac75ef3e6cdfbe1c105f179df08bb61224350a619390c7d0ae0a49c567bdd470ac75ef3e6cdfbe1c105f179df08bb61224350a619390c7d0ae0a49c567bdd47",
+      "0x2b04d6771aa832c8ac1d79a952ae8bacafaf093f023d4c83309995d0abe5e8b92b04d6771aa832c8ac1d79a952ae8bacafaf093f023d4c83309995d0abe5e8b9",
+      "0x1aa56b9fb75c97d91fa20c2ce17e86dc0c34b65836c9ba22107318b74634ec221aa56b9fb75c97d91fa20c2ce17e86dc0c34b65836c9ba22107318b74634ec22",
+      "0x1fc412ce46c3b2d344fec6237ac75033da3211884966dedd2b1a685b014f2d131fc412ce46c3b2d344fec6237ac75033da3211884966dedd2b1a685b014f2d13",
+      "0x29c27f9d279e3f52b51787e2a648a66f7a60e62497209bf986faa59419a91daa29c27f9d279e3f52b51787e2a648a66f7a60e62497209bf986faa59419a91daa",
+      "0x2e924806f112e278db1694edc2d6b7127053bcddfb24c55c9c47d578c47f0ec82e924806f112e278db1694edc2d6b7127053bcddfb24c55c9c47d578c47f0ec8",
+      "0x07cd5828f4e95899b5539065896b855b7684478e6cf2f9e1162ea9a03147184607cd5828f4e95899b5539065896b855b7684478e6cf2f9e1162ea9a031471846",
+      "0x173028dc3fc24d89b918ab4952f667ec2f8ea5341ce6c3202b0fefee6cf76041173028dc3fc24d89b918ab4952f667ec2f8ea5341ce6c3202b0fefee6cf76041",
+      "0x29a1ce46748dd1f268a52b64670d2dd170487b0eabfdf8e3280c52996af0356129a1ce46748dd1f268a52b64670d2dd170487b0eabfdf8e3280c52996af03561",
+      "0x0b070604db69fe26e6fff2c547a04321e96f99557a54427111d344a7f9c4fe212d87eaae6ee164a80a573baac4c0c747296fe504c130a06a78f63c488ca43cdf",
+      "0x0000000000000000000000000000000000000000000000000000000000000000",
+      "0x0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000127cf64d7b8255ebf08836e8c318db3ab29a0fcb45f0497e322acf1bfe5935b500c81f38fd033534fc4e483028f3f9066788d32f5a34de5b74b59aa9b5f01d0d6"
+    ]
+  }
+}
